refactor rule.tests & paragraph
