+++ src/rules/std.bench.ts
